+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneTrustedRootCertificateAndroidWork/MSFT_IntuneTrustedRootCertificateAndroidWork.psm1
@@ -276,17 +276,7 @@ function Set-TargetResource
             $AdditionalProperties['trustedRootCertificate'] = [Convert]::FromBase64String($AdditionalProperties['trustedRootCertificate'])
             Write-Verbose 'trustedRootCertificate converted to bytes.'
         }
-
         $CreateParameters.Remove('Id') | Out-Null
-
-        foreach ($key in ($CreateParameters.Clone()).Keys)
-        {
-            if ($CreateParameters[$key].GetType().Fullname -like '*CimInstance*')
-            {
-                $CreateParameters[$key] = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $CreateParameters[$key]
-            }
-        }
-
         $CreateParameters.Add('AdditionalProperties', $AdditionalProperties)
 
         #region resource generator code
@@ -323,16 +313,7 @@ function Set-TargetResource
             $AdditionalProperties['trustedRootCertificate'] = [Convert]::FromBase64String($AdditionalProperties['trustedRootCertificate'])
             Write-Verbose 'trustedRootCertificate converted to bytes.'
         }
-
         $UpdateParameters.Remove('Id') | Out-Null
-
-        foreach ($key in ($UpdateParameters.Clone()).Keys)
-        {
-            if ($UpdateParameters[$key].GetType().Fullname -like '*CimInstance*')
-            {
-                $UpdateParameters[$key] = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $UpdateParameters[$key]
-            }
-        }
         $UpdateParameters.Add('AdditionalProperties', $AdditionalProperties)
 
         #region resource generator code