+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneVPNConfigurationPolicyAndroidWork/MSFT_IntuneVPNConfigurationPolicyAndroidWork.psm1
@@ -474,16 +474,7 @@ function Set-TargetResource
                 $CreateParameters.Remove($keyName)
             }
         }
-
         $CreateParameters.Remove('Id') | Out-Null
-
-        foreach ($key in ($CreateParameters.Clone()).Keys)
-        {
-            if ($CreateParameters[$key].GetType().Fullname -like '*CimInstance*')
-            {
-                $CreateParameters[$key] = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $CreateParameters[$key]
-            }
-        }
 
         if ($AdditionalProperties.proxyServer)
         {
@@ -522,16 +513,7 @@ function Set-TargetResource
                 $UpdateParameters.Remove($keyName)
             }
         }
-
         $UpdateParameters.Remove('Id') | Out-Null
-
-        foreach ($key in ($UpdateParameters.Clone()).Keys)
-        {
-            if ($UpdateParameters[$key].GetType().Fullname -like '*CimInstance*')
-            {
-                $UpdateParameters[$key] = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $UpdateParameters[$key]
-            }
-        }
 
         if ($AdditionalProperties.proxyServer)
         {