+++ ResourceGenerator/M365DSCResourceGenerator.psm1
@@ -628,30 +628,12 @@ $($userDefinitionSettings.MOF -join "`r`n")
         `$createParameters = Rename-M365DSCCimInstanceParameter -Properties `$createParameters
         `$createParameters.Remove('Id') | Out-Null
 
-        `$keys = (([Hashtable]`$createParameters).Clone()).Keys
-        foreach (`$key in `$keys)
-        {
-            if (`$null -ne `$createParameters.`$key -and `$createParameters.`$key.GetType().Name -like '*CimInstance*')
-            {
-                `$createParameters.`$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject `$createParameters.`$key
-            }
-        }
 "@
         }
         $defaultUpdateParameters = @"
         `$updateParameters = ([Hashtable]`$boundParameters).Clone()
         `$updateParameters = Rename-M365DSCCimInstanceParameter -Properties `$updateParameters
-
         `$updateParameters.Remove('Id') | Out-Null
-
-        `$keys = (([Hashtable]`$updateParameters).Clone()).Keys
-        foreach (`$key in `$keys)
-        {
-            if (`$null -ne `$updateParameters.`$key -and `$updateParameters.`$key.GetType().Name -like '*CimInstance*')
-            {
-                `$updateParameters.`$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject `$updateParameters.`$key
-            }
-        }
 
 "@